2.0.4
Added (not fully completed) validation-catch behavior;
Removed unnecessary files and logic;
Worked on refactoring;
Started looking into integrating Auth0.
