+++ FleetJourney.Deploy/ApiGateway/apigateway-deployment.yaml
@@ -17,7 +17,7 @@ spec:
     spec:
       containers:
         - name: api-gateway
-          image: plaam/fleetjourney-api-gateway:1.0.9
+          image: plaam/fleetjourney-api-gateway:1.1.0
           imagePullPolicy: Always
           ports:
             - containerPort: 80
