+++ FleetJourney.Web/src/components/Sidebar.tsx
@@ -1,0 +1,361 @@
+import React, { useEffect, useRef, useState } from 'react';
+import { NavLink, useLocation } from 'react-router-dom';
+import SidebarLinkGroup from './SidebarLinkGroup.tsx';
+
+interface SidebarProps {
+    sidebarOpen: boolean;
+    setSidebarOpen: (arg: boolean) => void;
+}
+
+const Sidebar = ({ sidebarOpen, setSidebarOpen }: SidebarProps) => {
+    const location = useLocation();
+    const { pathname } = location;
+
+    const trigger = useRef<any>(null);
+    const sidebar = useRef<any>(null);
+
+    const storedSidebarExpanded = localStorage.getItem('sidebar-expanded');
+    const [sidebarExpanded, setSidebarExpanded] = useState(
+        storedSidebarExpanded === null ? false : storedSidebarExpanded === 'true'
+    );
+
+    useEffect(() => {
+        const clickHandler = ({ target }: MouseEvent) => {
+            if (!sidebar.current || !trigger.current) return;
+            if (
+                !sidebarOpen ||
+                sidebar.current.contains(target) ||
+                trigger.current.contains(target)
+            )
+                return;
+            setSidebarOpen(false);
+        };
+        document.addEventListener('click', clickHandler);
+        return () => document.removeEventListener('click', clickHandler);
+    });
+
+    useEffect(() => {
+        const keyHandler = ({ keyCode }: KeyboardEvent) => {
+            if (!sidebarOpen || keyCode !== 27) return;
+            setSidebarOpen(false);
+        };
+        document.addEventListener('keydown', keyHandler);
+        return () => document.removeEventListener('keydown', keyHandler);
+    });
+
+    useEffect(() => {
+        localStorage.setItem('sidebar-expanded', sidebarExpanded.toString());
+        if (sidebarExpanded) {
+            document.querySelector('body')?.classList.add('sidebar-expanded');
+        } else {
+            document.querySelector('body')?.classList.remove('sidebar-expanded');
+        }
+    }, [sidebarExpanded]);
+
+    return (
+        <aside
+            ref={sidebar}
+            className={`absolute left-0 top-0 z-9999 flex h-screen w-72.5 flex-col overflow-y-hidden bg-black duration-300 ease-linear dark:bg-boxdark lg:static lg:translate-x-0 ${
+                sidebarOpen ? 'translate-x-0' : '-translate-x-full'
+            }`}
+        >
+            <div className="flex items-center justify-between gap-2 px-6 py-5.5 lg:py-6.5">
+                <NavLink to="/">
+                    <h2 className="mb-4 ml-4 text-2xl font-semibold text-bodydark2">
+                        FleetJourney
+                    </h2>
+                </NavLink>
+            </div>
+
+            <div className="no-scrollbar flex flex-col overflow-y-auto duration-300 ease-linear">
+                {/* <!-- Sidebar Menu --> */}
+                <nav className="mt-5 py-4 px-4 lg:mt-9 lg:px-6">
+                    {/* <!-- Menu Group --> */}
+                    <div>
+                        <h3 className="mb-4 ml-4 text-sm font-semibold text-bodydark2">
+                            MENU
+                        </h3>
+                        <ul className="mb-6 flex flex-col gap-1.5">
+                            {/* <!-- Menu Item Dashboard --> */}
+                            <li>
+                                <NavLink
+                                    to="/"
+                                    className={`group relative flex items-center gap-2.5 rounded-sm py-2 px-4 font-medium text-bodydark1 duration-300 ease-in-out hover:bg-graydark dark:hover:bg-meta-4 ${
+                                        pathname.includes('/dashboard') &&
+                                        'bg-graydark dark:bg-meta-4'
+                                    }`}
+                                >
+                                    <svg
+                                        className="fill-current"
+                                        width="18"
+                                        height="19"
+                                        viewBox="0 0 18 19"
+                                        fill="none"
+                                        xmlns="http://www.w3.org/2000/svg"
+                                    >
+                                        <path
+                                            d="M6.10322 0.956299H2.53135C1.5751 0.956299 0.787598 1.7438 0.787598 2.70005V6.27192C0.787598 7.22817 1.5751 8.01567 2.53135 8.01567H6.10322C7.05947 8.01567 7.84697 7.22817 7.84697 6.27192V2.72817C7.8751 1.7438 7.0876 0.956299 6.10322 0.956299ZM6.60947 6.30005C6.60947 6.5813 6.38447 6.8063 6.10322 6.8063H2.53135C2.2501 6.8063 2.0251 6.5813 2.0251 6.30005V2.72817C2.0251 2.44692 2.2501 2.22192 2.53135 2.22192H6.10322C6.38447 2.22192 6.60947 2.44692 6.60947 2.72817V6.30005Z"
+                                            fill=""
+                                        />
+                                        <path
+                                            d="M15.4689 0.956299H11.8971C10.9408 0.956299 10.1533 1.7438 10.1533 2.70005V6.27192C10.1533 7.22817 10.9408 8.01567 11.8971 8.01567H15.4689C16.4252 8.01567 17.2127 7.22817 17.2127 6.27192V2.72817C17.2127 1.7438 16.4252 0.956299 15.4689 0.956299ZM15.9752 6.30005C15.9752 6.5813 15.7502 6.8063 15.4689 6.8063H11.8971C11.6158 6.8063 11.3908 6.5813 11.3908 6.30005V2.72817C11.3908 2.44692 11.6158 2.22192 11.8971 2.22192H15.4689C15.7502 2.22192 15.9752 2.44692 15.9752 2.72817V6.30005Z"
+                                            fill=""
+                                        />
+                                        <path
+                                            d="M6.10322 9.92822H2.53135C1.5751 9.92822 0.787598 10.7157 0.787598 11.672V15.2438C0.787598 16.2001 1.5751 16.9876 2.53135 16.9876H6.10322C7.05947 16.9876 7.84697 16.2001 7.84697 15.2438V11.7001C7.8751 10.7157 7.0876 9.92822 6.10322 9.92822ZM6.60947 15.272C6.60947 15.5532 6.38447 15.7782 6.10322 15.7782H2.53135C2.2501 15.7782 2.0251 15.5532 2.0251 15.272V11.7001C2.0251 11.4188 2.2501 11.1938 2.53135 11.1938H6.10322C6.38447 11.1938 6.60947 11.4188 6.60947 11.7001V15.272Z"
+                                            fill=""
+                                        />
+                                        <path
+                                            d="M15.4689 9.92822H11.8971C10.9408 9.92822 10.1533 10.7157 10.1533 11.672V15.2438C10.1533 16.2001 10.9408 16.9876 11.8971 16.9876H15.4689C16.4252 16.9876 17.2127 16.2001 17.2127 15.2438V11.7001C17.2127 10.7157 16.4252 9.92822 15.4689 9.92822ZM15.9752 15.272C15.9752 15.5532 15.7502 15.7782 15.4689 15.7782H11.8971C11.6158 15.7782 11.3908 15.5532 11.3908 15.272V11.7001C11.3908 11.4188 11.6158 11.1938 11.8971 11.1938H15.4689C15.7502 11.1938 15.9752 11.4188 15.9752 11.7001V15.272Z"
+                                            fill=""
+                                        />
+                                        <defs>
+                                            <clipPath id="clip0_130_9763">
+                                                <rect
+                                                    width="18"
+                                                    height="18"
+                                                    fill="white"
+                                                    transform="translate(0 0.052124)"
+                                                />
+                                            </clipPath>
+                                        </defs>
+                                    </svg>
+                                    Dashboard
+                                </NavLink>
+                            </li>
+                            {/* <!-- Menu Item Dashboard --> */}
+
+                            {/* <!-- Menu Item Employees --> */}
+                            <li>
+                                <NavLink
+                                    to="/employees"
+                                    className={`group relative flex items-center gap-2.5 rounded-sm py-2 px-4 font-medium text-bodydark1 duration-300 ease-in-out hover:bg-graydark dark:hover:bg-meta-4 ${
+                                        pathname.includes('employees') && 'bg-graydark dark:bg-meta-4'
+                                    }`}
+                                >
+                                    <svg
+                                        className="fill-current"
+                                        width="18"
+                                        height="19"
+                                        viewBox="0 0 18 19"
+                                        fill="none"
+                                        xmlns="http://www.w3.org/2000/svg"
+                                    >
+                                        <g clipPath="url(#clip0_130_9756)">
+                                            <path
+                                                d="M15.7501 0.55835H2.2501C1.29385 0.55835 0.506348 1.34585 0.506348 2.3021V15.8021C0.506348 16.7584 1.29385 17.574 2.27822 17.574H15.7782C16.7345 17.574 17.5501 16.7865 17.5501 15.8021V2.3021C17.522 1.34585 16.7063 0.55835 15.7501 0.55835ZM6.69385 10.599V6.4646H11.3063V10.5709H6.69385V10.599ZM11.3063 11.8646V16.3083H6.69385V11.8646H11.3063ZM1.77197 6.4646H5.45635V10.5709H1.77197V6.4646ZM12.572 6.4646H16.2563V10.5709H12.572V6.4646ZM2.2501 1.82397H15.7501C16.0313 1.82397 16.2563 2.04897 16.2563 2.33022V5.2271H1.77197V2.3021C1.77197 2.02085 1.96885 1.82397 2.2501 1.82397ZM1.77197 15.8021V11.8646H5.45635V16.3083H2.2501C1.96885 16.3083 1.77197 16.0834 1.77197 15.8021ZM15.7501 16.3083H12.572V11.8646H16.2563V15.8021C16.2563 16.0834 16.0313 16.3083 15.7501 16.3083Z"
+                                                fill=""
+                                            />
+                                        </g>
+                                        <defs>
+                                            <clipPath id="clip0_130_9756">
+                                                <rect
+                                                    width="18"
+                                                    height="18"
+                                                    fill="white"
+                                                    transform="translate(0 0.052124)"
+                                                />
+                                            </clipPath>
+                                        </defs>
+                                    </svg>
+                                    Employees
+                                </NavLink>
+                            </li>
+                            {/* <!-- Menu Item Employees --> */}
+
+                            {/* <!-- Menu Item Cars --> */}
+                            <li>
+                                <NavLink
+                                    to="/cars"
+                                    className={`group relative flex items-center gap-2.5 rounded-sm py-2 px-4 font-medium text-bodydark1 duration-300 ease-in-out hover:bg-graydark dark:hover:bg-meta-4 ${
+                                        pathname.includes('cars') && 'bg-graydark dark:bg-meta-4'
+                                    }`}
+                                >
+                                    <svg
+                                        className="fill-current"
+                                        width="18"
+                                        height="19"
+                                        viewBox="0 0 18 19"
+                                        fill="none"
+                                        xmlns="http://www.w3.org/2000/svg"
+                                    >
+                                        <g clipPath="url(#clip0_130_9756)">
+                                            <path
+                                                d="M15.7501 0.55835H2.2501C1.29385 0.55835 0.506348 1.34585 0.506348 2.3021V15.8021C0.506348 16.7584 1.29385 17.574 2.27822 17.574H15.7782C16.7345 17.574 17.5501 16.7865 17.5501 15.8021V2.3021C17.522 1.34585 16.7063 0.55835 15.7501 0.55835ZM6.69385 10.599V6.4646H11.3063V10.5709H6.69385V10.599ZM11.3063 11.8646V16.3083H6.69385V11.8646H11.3063ZM1.77197 6.4646H5.45635V10.5709H1.77197V6.4646ZM12.572 6.4646H16.2563V10.5709H12.572V6.4646ZM2.2501 1.82397H15.7501C16.0313 1.82397 16.2563 2.04897 16.2563 2.33022V5.2271H1.77197V2.3021C1.77197 2.02085 1.96885 1.82397 2.2501 1.82397ZM1.77197 15.8021V11.8646H5.45635V16.3083H2.2501C1.96885 16.3083 1.77197 16.0834 1.77197 15.8021ZM15.7501 16.3083H12.572V11.8646H16.2563V15.8021C16.2563 16.0834 16.0313 16.3083 15.7501 16.3083Z"
+                                                fill=""
+                                            />
+                                        </g>
+                                        <defs>
+                                            <clipPath id="clip0_130_9756">
+                                                <rect
+                                                    width="18"
+                                                    height="18"
+                                                    fill="white"
+                                                    transform="translate(0 0.052124)"
+                                                />
+                                            </clipPath>
+                                        </defs>
+                                    </svg>
+                                    Cars
+                                </NavLink>
+                            </li>
+                            {/* <!-- Menu Item Cars --> */}
+
+                            {/* <!-- Menu Item Trips --> */}
+                            <li>
+                                <NavLink
+                                    to="/trips"
+                                    className={`group relative flex items-center gap-2.5 rounded-sm py-2 px-4 font-medium text-bodydark1 duration-300 ease-in-out hover:bg-graydark dark:hover:bg-meta-4 ${
+                                        pathname.includes('trips') && 'bg-graydark dark:bg-meta-4'
+                                    }`}
+                                >
+                                    <svg
+                                        className="fill-current"
+                                        width="18"
+                                        height="19"
+                                        viewBox="0 0 18 19"
+                                        fill="none"
+                                        xmlns="http://www.w3.org/2000/svg"
+                                    >
+                                        <g clipPath="url(#clip0_130_9756)">
+                                            <path
+                                                d="M15.7501 0.55835H2.2501C1.29385 0.55835 0.506348 1.34585 0.506348 2.3021V15.8021C0.506348 16.7584 1.29385 17.574 2.27822 17.574H15.7782C16.7345 17.574 17.5501 16.7865 17.5501 15.8021V2.3021C17.522 1.34585 16.7063 0.55835 15.7501 0.55835ZM6.69385 10.599V6.4646H11.3063V10.5709H6.69385V10.599ZM11.3063 11.8646V16.3083H6.69385V11.8646H11.3063ZM1.77197 6.4646H5.45635V10.5709H1.77197V6.4646ZM12.572 6.4646H16.2563V10.5709H12.572V6.4646ZM2.2501 1.82397H15.7501C16.0313 1.82397 16.2563 2.04897 16.2563 2.33022V5.2271H1.77197V2.3021C1.77197 2.02085 1.96885 1.82397 2.2501 1.82397ZM1.77197 15.8021V11.8646H5.45635V16.3083H2.2501C1.96885 16.3083 1.77197 16.0834 1.77197 15.8021ZM15.7501 16.3083H12.572V11.8646H16.2563V15.8021C16.2563 16.0834 16.0313 16.3083 15.7501 16.3083Z"
+                                                fill=""
+                                            />
+                                        </g>
+                                        <defs>
+                                            <clipPath id="clip0_130_9756">
+                                                <rect
+                                                    width="18"
+                                                    height="18"
+                                                    fill="white"
+                                                    transform="translate(0 0.052124)"
+                                                />
+                                            </clipPath>
+                                        </defs>
+                                    </svg>
+                                    Trips
+                                </NavLink>
+                            </li>
+                            {/* <!-- Menu Item Trips --> */}
+
+                            {/* <!-- Menu Item Settings --> */}
+                            <li>
+                                {/* <!-- Menu Item Auth Pages --> */}
+                                <SidebarLinkGroup
+                                    activeCondition={
+                                        pathname === '/auth' || pathname.includes('auth')
+                                    }
+                                >
+                                    {(handleClick, open) => {
+                                        return (
+                                            <React.Fragment>
+                                                <NavLink
+                                                    to="#"
+                                                    className={`group relative flex items-center gap-2.5 rounded-sm py-2 px-4 font-medium text-bodydark1 duration-300 ease-in-out hover:bg-graydark dark:hover:bg-meta-4 ${
+                                                        (pathname === '/auth' || pathname.includes('auth')) &&
+                                                        'bg-graydark dark:bg-meta-4'
+                                                    }`}
+                                                    onClick={(e) => {
+                                                        e.preventDefault();
+                                                        sidebarExpanded
+                                                            ? handleClick()
+                                                            : setSidebarExpanded(true);
+                                                    }}
+                                                >
+                                                    <svg
+                                                        className="fill-current"
+                                                        width="18"
+                                                        height="19"
+                                                        viewBox="0 0 18 19"
+                                                        fill="none"
+                                                        xmlns="http://www.w3.org/2000/svg"
+                                                    >
+                                                        <g clipPath="url(#clip0_130_9814)">
+                                                            <path
+                                                                d="M12.7127 0.55835H9.53457C8.80332 0.55835 8.18457 1.1771 8.18457 1.90835V3.84897C8.18457 4.18647 8.46582 4.46772 8.80332 4.46772C9.14082 4.46772 9.45019 4.18647 9.45019 3.84897V1.88022C9.45019 1.82397 9.47832 1.79585 9.53457 1.79585H12.7127C13.3877 1.79585 13.9221 2.33022 13.9221 3.00522V15.0709C13.9221 15.7459 13.3877 16.2802 12.7127 16.2802H9.53457C9.47832 16.2802 9.45019 16.2521 9.45019 16.1959V14.2552C9.45019 13.9177 9.16894 13.6365 8.80332 13.6365C8.43769 13.6365 8.18457 13.9177 8.18457 14.2552V16.1959C8.18457 16.9271 8.80332 17.5459 9.53457 17.5459H12.7127C14.0908 17.5459 15.1877 16.4209 15.1877 15.0709V3.03335C15.1877 1.65522 14.0627 0.55835 12.7127 0.55835Z"
+                                                                fill=""
+                                                            />
+                                                            <path
+                                                                d="M10.4346 8.60205L7.62207 5.7333C7.36895 5.48018 6.97519 5.48018 6.72207 5.7333C6.46895 5.98643 6.46895 6.38018 6.72207 6.6333L8.46582 8.40518H3.45957C3.12207 8.40518 2.84082 8.68643 2.84082 9.02393C2.84082 9.36143 3.12207 9.64268 3.45957 9.64268H8.49395L6.72207 11.4427C6.46895 11.6958 6.46895 12.0896 6.72207 12.3427C6.83457 12.4552 7.00332 12.5114 7.17207 12.5114C7.34082 12.5114 7.50957 12.4552 7.62207 12.3145L10.4346 9.4458C10.6877 9.24893 10.6877 8.85518 10.4346 8.60205Z"
+                                                                fill=""
+                                                            />
+                                                        </g>
+                                                        <defs>
+                                                            <clipPath id="clip0_130_9814">
+                                                                <rect
+                                                                    width="18"
+                                                                    height="18"
+                                                                    fill="white"
+                                                                    transform="translate(0 0.052124)"
+                                                                />
+                                                            </clipPath>
+                                                        </defs>
+                                                    </svg>
+                                                    Authentication
+                                                    <svg
+                                                        className={`absolute right-4 top-1/2 -translate-y-1/2 fill-current ${
+                                                            open && 'rotate-180'
+                                                        }`}
+                                                        width="20"
+                                                        height="20"
+                                                        viewBox="0 0 20 20"
+                                                        fill="none"
+                                                        xmlns="http://www.w3.org/2000/svg"
+                                                    >
+                                                        <path
+                                                            fillRule="evenodd"
+                                                            clipRule="evenodd"
+                                                            d="M4.41107 6.9107C4.73651 6.58527 5.26414 6.58527 5.58958 6.9107L10.0003 11.3214L14.4111 6.91071C14.7365 6.58527 15.2641 6.58527 15.5896 6.91071C15.915 7.23614 15.915 7.76378 15.5896 8.08922L10.5896 13.0892C10.2641 13.4147 9.73651 13.4147 9.41107 13.0892L4.41107 8.08922C4.08563 7.76378 4.08563 7.23614 4.41107 6.9107Z"
+                                                            fill=""
+                                                        />
+                                                    </svg>
+                                                </NavLink>
+                                                {/* <!-- Dropdown Menu Start --> */}
+                                                <div
+                                                    className={`translate transform overflow-hidden ${
+                                                        !open && 'hidden'
+                                                    }`}
+                                                >
+                                                    <ul className="mt-4 mb-5.5 flex flex-col gap-2.5 pl-6">
+                                                        <li>
+                                                            <NavLink
+                                                                to="/auth/signin"
+                                                                className={({ isActive }) =>
+                                                                    'group relative flex items-center gap-2.5 rounded-md px-4 font-medium text-bodydark2 duration-300 ease-in-out hover:text-white ' +
+                                                                    (isActive && '!text-white')
+                                                                }
+                                                            >
+                                                                Sign In
+                                                            </NavLink>
+                                                        </li>
+                                                        <li>
+                                                            <NavLink
+                                                                to="/auth/signup"
+                                                                className={({ isActive }) =>
+                                                                    'group relative flex items-center gap-2.5 rounded-md px-4 font-medium text-bodydark2 duration-300 ease-in-out hover:text-white ' +
+                                                                    (isActive && '!text-white')
+                                                                }
+                                                            >
+                                                                Sign Up
+                                                            </NavLink>
+                                                        </li>
+                                                    </ul>
+                                                </div>
+                                                {/* <!-- Dropdown Menu End --> */}
+                                            </React.Fragment>
+                                        );
+                                    }}
+                                </SidebarLinkGroup>
+                                {/* <!-- Menu Item Auth Pages --> */}
+                            </li>
+                        </ul>
+                    </div>
+                </nav>
+                {/* <!-- Sidebar Menu --> */}
+            </div>
+        </aside>
+    );
+};
+
+export default Sidebar;
