+++ FleetJourney.Web/src/pages/Dashboard.tsx
@@ -1,4 +1,4 @@
-import DefaultLayout from '../../layout/DefaultLayout.tsx';
+import DefaultLayout from '../layout/DefaultLayout.tsx';
 
 const Dashboard = () => {
     return (
@@ -6,7 +6,7 @@ const Dashboard = () => {
             <div className="mt-4 grid grid-cols-12 gap-4 md:mt-2 md:gap-6 2xl:mt-7.5 2xl:gap-7.5">
                 <div className="col-span-10 xl:col-span-12">
                     <p className="text-4xl text-fail font-semibold">
-                        TO BE IMPLEMENTED
+                        ?
                     </p>
                 </div>
             </div>
