+++ FleetJourney.Web/vite.config.js
@@ -1,6 +1,9 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
-export default defineConfig({
-  plugins: [react()],
+export default defineConfig ({
+    plugins: [react()],
+    preview: {
+        port: 4040
+    },
 })
